Allow dashboard creation only permitted users
Before this PR, a user with only stream reading permissions
was able to create a dashboard by clicking on the displayed
dashboard button.

With this change, we restrict the creation of dashboards
on the backend and hide the dashboard create button
if the user does not have the needed permissions.
Additionally we only render the dashboards/new page
if the user has the needed permissions.

Fixes #7729