Removing n+1's in `StreamService#loadAll` by consolidating external loads
Before this change, calling `StreamService#loadAll` triggered loading of
associated outputs and index sets, _both_ resulting in n+1 each.

This change is now placing fetching of outputs and index sets outside
the loop iterating over each stream, leaving only simple lookups to be
done inside the loop, effectively turning this into a call with a
_constant_ number of mongodb queries.

Simple tests on my setup showed an improvement of ~270ms -> ~110ms.
While this is already impressive, I expect a much bigger improvement for
setups with a high number of outputs.