Replace mutating state inline in stores with new object creation. (#7864)
In #7671, a `shouldComponentUpdate` function for `connect` was
implemented, that prevents a rerendering of its children when the state
supplied by a store did not change. The implication of this is that a
store should always propagate a new object when its state has changed,
instead of mutating its state inline. In the latter case, the consumer
of the state will hold a reference to the previous state that is the
same as the new state, so it will appear to be equal. When implementing
that PR, a search was performed for patterns where a store was assigning
a new value to an instance variable (anything that looks like
`this(\.\w+)+ = .+;`) and patterns like this were changed to create a
new object that is assigned to the instance variable.

Unfortunately patterns where the bracket notation (e.g. `this.foo[bar] =
42`) were left out, leading to some inline mutations being left in
stores. This PR is addressing these. To find code positions where this
is necessary, a regex search for `this(\.\w+)+\[.+\] =` was performed
and the affected code parts where fixed.