Replacing usage of `Object.fromEntries` with `reduce`.
In #7864 a usage of `Object.fromEntries` was introduced. Unfortunately
this causes tests to fail on node v10. Luckily this was caught by
@linuspahl. My attempts to configure babel to include the required
transforms for this (by setting `{ targets: { node: true } }`) did not
succeed, but cause other errors. Therefore I decided to replace the
usage with a call to `reduce`, which is less readable, but has the same
outcome.