Do not filter querys with empty search types
Prior to this change, we filtered the queries with
empty search types. Without this any queries
(which can happen on an completly new dashboard)
the resulting response to the frontend is empty.
Which leads to an loading spinner.

This change will remove the filtering of the queries.
So a search from a search and from a dashboard
are producing the same result.

This is far from idal since we now produce a
empty result set which is not needed. But this
is the safest option for now.

Fixes #7807