Enforce Unix newlines for Ace editor
If documents in the Ace editor are modified on a Windows browser,
it would result in windows style newlines, which Graylog doesn't handle
correctly.

The Ace editor defaults to 'auto' mode, which switches to
windows mode, if the content is pasted on Windows.

It's better to uniformly have unix style newlines everywhere,
thus enforce unix mode as a default for Ace.

Fixes #7889
Fixes https://github.com/Graylog2/collector-sidecar/issues/389