Add decorators to ExportMessagesCommand
Finding the decorators to apply belongs in CommandFactory. This will also make it simpler to add in support from decorators that are not directly attached to message lists.
E.g., decorators configured for streams, if we want to use them, or decorators a user might want to specify on API requests.