Work around non-existing `Event` constructor in IE11.
In Internet Explorer 11, calling `new Event()` raises an exception.
Instead, a new event needs to be created using
`el.createEvent`/`event.initEvent`, the latter being deprecated on all
other major browsers.

This PR is abstracting the uglyness of this circumstance by offering a
`createEvent` function which returns an event after checking if the
`Event` constructor is available, using the other approach if not.

Fixes #7670.