Fix assignment of default timerange for export requests
Default time range was originally a `RelativeRange`. When it was converted to an `AbsoluteRange` it should not have been assigned to a static field any more, because the correct absolute time range must be computed dynamically the moment the request is built.
Instead it was computed once and remained statically set to the original "last five minutes" for all subsequent requests. This commit fixes that.