Allow writing cluster config object with custom key.
Previously, the `ClusterConfigService` interface allowed writing a
cluster config object by passing it and extracting the key/type it is
stored as from its class name. This is fine for most use cases, but
makes writing migrations that migrate the structure of a pre-existing
cluster config object non-trivial if separate classes are used for
stability. Those result in different class names, therefore being
persisted with different types than the cluster config object that
should be migrated.

This change is adding a `write(String key, Class<T> payload)` method to
the interface, which allows specifying a custom key for the payload,
therefore making it independent of the class implementing it.