Reset job trigger on EventDefinition updates
This replaces the former approach to fixing
the job schedules upon updates.

The previous approach only worked on job triggers
that haven't been modified by the scheduler.

Instead on trying to find a usable timerange from the
old job trigger, we simply wipe the job trigger data,
and start over with a timerange ending on the current time.
This timerange comes from the updated jobDefinition,
which is recreated upon the update.
This works because EventProcessorExecutionJob.execute()
can also take the parameters from the JobDefintion if the
job trigger data is not present.

Fixes #7841