Don't enable CORS by default
Having a default of allowing any Origin is not a sane security default.

In a production environment, all components of Graylog (frontend,
assets, API) will be served behind a single Origin.
That's why CORS shouldn't be needed for most users.

For development or other special cases, this now needs to be enabled explicitly.

While here, remove some duplicates in the CORSFilter and simply use
a wildcard instead of responding with the Origin that made the query,
which is effectively the same.