Allow arbitrary key sizes for AESTools.encrypt/decrypt. (3.2) (#8098)
* Allow arbitrary key sizes for AESTools.encrypt/decrypt.

In the current state, the `AESTools.encrypt/decrypt` methods rely on key
sizes being 16, 24 or 32 bytes. This makes functionality using it with
user-supplied keys fail if their lengths are not exactly that size.

In order to allow arbitrary key sizes, this change is cutting keys if
they are longer than 32 bytes or pads them if they are shorter than the
desired key size. Also, it uses the first of the three key sizes which
can hold as most of the supplied key as possible.

* Small refactorings.

* Adding asserts and javadoc.

* Properly handle unicode strings.

* Fixing rebase conflicts.