Throw exception when creating session but auth service is not available
When an authenticating realm is using an external service, there was no
way to indicate that authentication failed in case the external service
was unavailable at the moment of checking the credentials. It could only
indicate that authentication was unsuccessful in general.
Therefore, in the unavailability case, the user would probably assume
that the credentials were wrong.

With this change, an authenticating realm may throw an
`AuthenticationServiceUnavailableException` which will be retained if
none of the realms can successfully authenticate the user. This will
then be indicated to the API client with a non-successful status code
and response message.