Create separate module for ES-specific implementations.
This change is the first in a series aiming to separate our
Elasticsearch-specific implementation code from its consumers. It is
creating a separate maven module that is holding our current
implementation for:

  * Indexing Messages
  * Searching/aggregation previously indexed Messages
  * Index Management
  * Cluster health checks and maintenance

The goal of this is to encapsulate and hide version-specific
implementation details, allowing to switch the implementation code in
the future.