Include storage module in server assembly.
With the introduction of the separate module holding storage-related
implementation code, we need to explicitly include it in the server
assembly. The reasoning is:

  * The `graylog-storage-elasticsearch6` module is depending on the `graylog2-server` module, but not the other way around (to prevent a circular dependency)
  * Due to the `graylog2-server` module not knowing about the `graylog-storage-elasticsearch6` module, it is not included in its jar
  * As the `graylog-storage-elasticsearch6` is implemented using our plugin architecture (using the `ServiceLoader`), it should be loaded just as other plugins

Therefore we are building the storage module as a separate jar and
include it in the assembly (which results in a tarball).

This also means that other parts which require the storage module at
runtime need to make it available to the server as a plugin (e.g. this
affects the `full-backend-tests` module, a PR including the required
changes for it to load the storage module plugin will follow).