Create `distribution` module stub for assembly.
This change is creating a `distribution` maven module, which is solely
used for the final assembly step creating the distribution tarball for
the server. This is required to ensure that all module's jar files have
been built before. Before this change, the final assembly was part of
the `graylog2-server` module, so it was not possible to reference
artifacts from modules built later in the chain.