Include storage module in Graylog setup of integration tests.
As we are now building a separate storage module that is holding
implementations for interfaces encapsulating ES-specific functionality,
we need to ensure that this module is present at runtime. This is
relevant for our dockerized integration tests as well.

This PR is ensuring that the storage module is mounted into the docker
container into the plugin directory so it is loaded during startup.