Preventing exception or indexing without document id when moving.
Before this change, when moving documents from one index to another, it
could theoretically happen that a document either has no source field or
no id. In the first case an exception would be thrown and further moving
aborted, in the other case a document without an id would be attempted
to index.

This small change is handling the optionality of the source field
properly and only adds a bulk indexing action to the request if both
source and id fields exist.