Extract interface from `Cluster` class. (#8233)
* Extract interface from `Cluster` class.

This change is extracting a `ClusterAdapter` interface from the `Cluster` class containing ES-specific implementation details and creates an implementation class (`ClusterAdapterES6`) containing the current code.

During the extraction most of the signatures were kept intact, besides those which use the ES `Health.Status` class. It was replaced with a very simple own `HealthStatus` enum which is mappable 1:1 with the current implementation of `Health.Status`. It might make sense to move away from colors to more precise identifiers (e.g. "Healthy", "Degraded", "Unhealthy"). Due to the change of the return type of those methods it was also required to change some of the logic in the consumers of these classes, but it was attempted to keep those to a minimum in favor of later refactorings.

* Incorporating changes from #8050.