Mark legacy aggregation API as deprecated.
In `3.2.0`, the Views API was introduced, offering extensive aggregation
capabilities. Since then, the previous aggregation API was not used by
the frontend anymore.

To free up time spent on maintaining another legacy API, those endpoints
are marked as deprecated and are removed in `4.0`, offering Graylog core
developers the chance to invest time on other parts of the product.

The deprecated endpoints include:

- `/search/universal/(absolute|relative|keyword)/`
    - `terms-histogram`
    - `histogram`
    - `fieldhistogram`
    - `stats`
    - `termsstats`
    - `terms`
- `/sources`

As well as the following methods from the `Searches` class:

- `fieldHistogram`
- `histogram`
- `terms`
- `termsHistogram`
- `termsStats`