Move widget to tab (#8236)
* Add move widget to tab functionallity

Prior to this change, there was no code providing the possibility
to move a widget from one tab to another.

This change will add a function.

* Create MoveWidgetToTab Modal

Prior to this change, there was no Modal where a user
could select to which tab he would like to move a widget.

This change will create a new Modal and connects it
to the former MoveWidgetToTab logic.

* Fix failing snapshot tests

Prior to these changes, we did not update the search id,
which needed to be change, since we changed the search.

This change will mock the id generation so we do not have
failing snapshots tests with every run.

* Make it possible to keep a copy of a widget

Prior to this change, you could only move a widget
to a different page. Now you can also copy it instead.

This change will add a checkbox where the user can
decide to copy the widget instead of moving it.

* Rename MoveWidgetToTab to MoveWidgetToTabModal

* Use BootstrapModalForm and beautify ListGroupItem usage

* Filter active query from selection list

* Use QueryTitle for title generation

* Reuse existing types instead of reinventing

* use better VariableName and throw Error when type is unexpected

* When moving a widget also add new WidgetPostion

* fix linter warnings

* Use QueryIds store to keep the order of pages

* Disable submit button if nothing is selected

* Use the first row notation from WidgetGrid (starting by 1)

* zero was correct

* Recalcutate WidgetPositions

Prior to this change, we simply add a new WidgetPosition with row 1 and
col 1. This will only lead to the problem that now 2 widgets have this
coordinates.

This change will reuse the calculation method from
CurrentViewStateStore.

* Also move the title of the widget

* useCallback to memorize the callbacks