Reuse `Messages#bulkIndex` for event indexing.
This change was targeted to reimplement the `EventIndexAdapter` for ES7.  It was noticed early on that the event indexing code did not implement retrying on IOExceptions or for read-only indices/cluster blocks yet, also it does not handle indexing errors as well as the regular bulk indexing code does. The reason for its existence was a different preparation of event timestamps compared to indexing regular messages.

Instead of reimplementing the event indexing code, this change does now target at reusing the existing bulk indexing code by implementing a `Indexable` interface that both `Message` and `Event` are supposed to implement/extend. It collects all required methods to index a message or event. This allows us to reuse `Messages#bulkIndex(*)` for both indexing cases, making retrying and error handling logic consistent.