Fix grantee selection logic
If there is no "selected_grantee_capabilities" in the share request,
return the active shares as selected.
If there is an empty "selected_grantee_capabilities" in the share
request, it means all shares should be removed.

Also merge EntitySharePrepareRequest and EntityShareRequest into one
object and also rename EntitySharePrepareResponse to
EntityShareResponse.