Fix flow type errors
Prior to this change, flow type did remark my priror changes
to the defaultPayload.

This change is removing the exact object type since the object
can now be empty. Since the payload from update and prepare are
the same, I made one payload type out of both.

Flow still had problems recognizing the type of the payload
when using the function, so I cast the type.