Introduce UserContext helper
This allows us to access the current user context
from anywhere, whithout having to pass it through from the
calling RestResource.

Shiro binds the current user to a ThreadContext, which is
accessible from anywhere.

We also provide a UserContext#runAs() method,
that can be used to create a temporary user context
to run code with.