Use correct ES exception class in storage adapter. (#8515)
This change is effectively reverting ffecb94e. What seemed to be a good idea then (returning our wrapped exception earlier, from the client) turned out to be breaking the 413 handling as we expect to have access to the raw exception coming from the client there.

Therefore this commit is reverting this change in order to restore functionality, a future change could improve throwing a wrapped exception early again, after fixing its consumers.