Implement `ClusterAdapter` for ES7. (#8459)
* Starting and binding adapter.

* Implement `ClusterAdapter` for ES7.

This change is implementing a `ClusterAdapter` for ES7. It provides
cluster-related functionality and is consumed mainly by the `Cluster`
class.

During the implementation of this using the Elasticsearch high-level
REST client, it became obvious that it is not supporting API endpoints
that Jest supported, namely the `_cat` and the `_nodes` endpoints. For
the `_cat` API, a `CatApi` class was implemented, that returns proper
value objects for the responses. For the other missing APIs it was
different, as only a small number of keys were used from the responses,
with little overlap. Therefore a `PlainJsonApi` class created, that
allows lightweight querying of single endpoints, returning the plain
JSON structure.

* Cleaning up unused class.

* Moving duplicated `statsForIndices` method to `Cluster`, exposing its used methods in adapter interface.

* Removing unneeded `deflectorHealth` from adapter, reusing `health` in `Cluster`.

* Reusing `HealthStatus.fromString` in `ClusterAdapterES6`.

* Moving test method to superclass, removing comment, improving name instead.

* Fixing method signature.

* Delegating to `Cluster` instead of using `ClusterAdapter` in `ClusterStatsService`.

* Pulling in `ElasticsearchClient` dependency into `PlainJsonApi` & `CatApi`.

* Removing obsolete dependency from `ClusterStatsService`.