type check unary NOT operator
the type checker did not properly check the sub expression type for the NOT operator, which could lead to runtime errors
NOT binds closer than comparison (==), which is a common source of mistakes
in this case a string valued function as being put into `NOT func() == "string"` which parses as `(NOT func()) == "string"`
the type checker did not pick up that `func()` wasn't boolean valued and the interpreter failed at runtime due to a classcast exception.

after this change the type checker requires the expression passed to NOT to be of boolean type and makes this a parse-time error

this should not have any incompatibility issues, because it never executed successfully (unless the expression was short-circuited). in any case, rules like this are severely broken and need to be fixed by the user