Unconditionally include time range filter in created search for ES7. (#8606)
* Unconditionally include time range filter in created search for ES7.

Before this change, a subtle bug was introduced by wrong optional
chaining. The effect was that the time range was included in a created
search only if the filter query string is present and not a wildcard.

This change is fixing that by streamlining filter creation and addition
to the search source builder.

* Adding test helper for JSONPath assertions.

* Adding test for time range inclusion in seach request.

* Using absolute instead of relative range in test.