Expose GRNPermissions to frontend
The frontend needs to know which entities a user
currently owns. This is expressed with a new Permissions
type, that is incompatible with the existing WildcardPermission.

Introduce a new field `grn_permissions` into the UserSummary object.

The serialized format of a GRNPermission is

`<permission>|<grn>`

e.g.: `entity:own|grn::::dashboard:5e2afc66cd19517ec2dabaf3`

Fixes #8603