Add function to PaginatedUserService to load users by username
Prior to this change, there was no way to retrive a paginated list of
users by there username. This is needed for the Team Details Page
where we want to display the Users of Team.

Also we add `name` and `description` to UserOverview, which is simply
giving `username` and `fullname`.

This change will also fix everloading UserDetailsPage and a small typo.