Remove completed marker, if present after ES downgrade
In case a customer goes back to pre-7 Elasticsearch version after the migration has been run, we remove the completed marker. This makes sure that the migration would run again once the customer goes back to a version >= 7.
Also added an integration test to make sure that adding the field aliases is idempotent.