Enable Search After for ES7 exports by adding a field alias for gl2_message_id in event indices (#8672)
* Remove tie breaker workarounds for ES7 SearchAfter

* Use SearchAfter for ES7 export

* Remove ES7 export scrolling

* Added new migration for adding the field alias

- Created migration class with basic functionality
- Created tests for basic functionality

* Implement Elasticsearch adapter for new migration

Only an ES7 adapter was added. Since field aliases were introduced for 6.4, adding in ES6 support would be too complicated to be worth it.

- Created ES7 adapter for new migration
- Created integration test for new adapter
- Allow accessing the RestHighLevelClient from ElasticsearchInstanceES7

* Skip migration for ES versions < 7

Since field aliases were only introduced for 6.4, we will only use them for version 7 and above. Otherwise we would have to differentiate between different minor versions for 6, which would make things considerably more complicated.

* Wire up Guice bindings for migration

Also added an ES6 implementation for the adapter. It should never be called and will throw an exception otherwise. This makes sure that:
 - We don't accidentally add field aliases for ES versions < 6.4 where they are not supported
 - The migration doesn't mark itself as completed without actually adding field aliases.

 When a customer runs Graylog 4.0 on ES6, field aliases will not be added, but the migration will try again on subsequent server restarts. So if they upgrade to ES7 at a later point and restart the server, the migration will run and add the field aliases.

* Added field alias to EventsIndexMapping7

* Use test client instead of IndicesAdapterES7

...for creating test indices in V20200730000000_AddGl2MessageIdFieldAliasForEventsES7IT

* Fix formatting in VersionAwareStorageModule

* Remove completed marker, if present after ES downgrade

In case a customer goes back to pre-7 Elasticsearch version after the migration has been run, we remove the completed marker. This makes sure that the migration would run again once the customer goes back to a version >= 7.
Also added an integration test to make sure that adding the field aliases is idempotent.

* Fixed accidentally introduced star imports