Migrate streams and dashboards roles to grants
This migrates the currently UI configurable roles
for dashboards and streams into grants.

If the permissions of a role are empty after the migration, it gets removed
completely.

If entities has permissions other than read and edit, we don't migrate
it, so we don't end up having permissions on two different places for a
single entity.