Consolidate index mappings and associated tests (#8632)
* Parameterize unsupported version tests in IndexMappingFactoryTest

* Add tests for ES7 to IndexMappingFactoryTest

* Convert EventsIndexMappingTest to JUnit 5

* Parameterize tests in EventsIndexMappingTest

* Include ES7 in EventsIndexMappingTest#createsValidMappingTemplates

* Extract base class EventsIndexMapping

The mapping is the same except for the "message" key that's only necessary for ES6

* Added IndexMapping7Test

* Move tests for unsupported ES version for event mappings

The tests are equivalent to those for message mappings which were already present in IndexMappingFactoryTest.

* Consolidated index mapping tests in one class

- Added IndexMappingTest
- Consolidated tests from IndexMapping5Test, IndexMapping6Test and IndexMapping7Test into one parameterized test in IndexMappingTest.
- Deleted IndexMappingTestBase and moved contents into IndexMappingTest

This has the added benefit of being more similar to equivalent tests in EventsIndexMappingTest. The difference of using fixture files for the message mapping tests was left in place. It makes sense to use fixtures here, because the mappings are somewhat more complicated than those for events.

* Use explicit LOCALE in String#format call

* Removed old TODO comment

* Added field alias to EventsIndexMapping7