Fix publicPath for file-loader rules in plugin webpack configs
Assets like SVG or WOFF files are accessible at the default "/assets"
path. This works for files in the server core but doesn't work for files
in plugins.

For plugins the path needs to be "/assets/plugin/<plugin-name>" to make
sure the web server serves the files from the plugin JAR instead of the
server JAR.

This affects .svg, .woff, .woff2 and .ttf files because those are using
the "file-loader" webpack loader. Images like .png, .gif and .jpg are
using the "url-loader", not sure if they are affected by that bug as
well.