Implement HasOwnership which checks for sharing
## Motivation
Prior to this change, we always show the entity sharing modal button
even if you were not allowed to see it which would lead into a error
screen.

## Description
This change will implement a switching component which decides if a
child will be rendered depending on if the user has ownership over the
entity passed on.