Removing tests snapshotting layout excessively.
This change is removing snapshots containing too much layout, which is
not part of the system under test. Those are mostly tests which render a
component and throw all of the returned markup against `toMatchSnapshot`,
resulting in huge snapshots, which do not provide us a benefit going
beyond (ideally) a check that the component renders at all.

The additional downsides of these are:

  * Polluting PRs which change theming or central components, easily
blowing up n*10 loc changes to n*10000 loc changes, making a proper code
review impossible and effectively shadowing the actual changes.
  * Due to the size of the snapshots, test negatives are dismissed too easily
during development. Instead of investigating big diffs, snapshots are
just updated, rendering the mechanism ineffective.
  * The existence of snapshot tests suggest (falsely) proper coverage of
the SUT. Instead, behavior-focused tests should be used.

This should not suggest that snapshot tests should not be used _at all_.
They are useful for cases where isolated, stable results are checked.
This includes functions operating on bigger data structures, where
generating test expectations using snapshots are feasible and the
results are stable.