Serialize body of new messages using object mapper for ES7.
This change is using the configured system-wide `ObjectMapper` instance
to serialize the body of new messages supposed to be bulk indexed by
Elasticsearch when creating the request. This change is required for
ES7, as it does not allow configuring a custom `ObjectMapper` instance
to be used for serialization, in contrast to `Jest` which is used for
ES6.