Fix invalid options for force merge request during index optimization.
This change removes the "only expunge deletes" option from the force
merge request, as it contradicts with the "max num segments" option. For
every execution of the request, a warning is logged about this, also
indicating that while the request is still performed, it might not in
future ES versions.