Add share button to Dashboard LIst (#8830)
## Motivation
Prior to this change, the user needed to go to the Dashboard view/edit
page to add collaborators to a dashboard.

## Description
This change is now adding a MenuItem to the dashboard list so the user
can share the dashboard from the over view list. This might be helpful
if a user needs to share a lot of dashboards.