Use correct interval type for time bucketing.
The Elasticsearch high-level REST client offers two ways to define the
interval for time-based bucketing, fixed and calendar intervals. Those
can be used for different interval specifications. Before this change,
the fixed interval was used for everything, leading to exceptions being
thrown e.g. when the interval was specified as `1M` (one month).

Therefore, this change checks if the given interval matches the ones
suitable for calendar intervals or uses fixed intervals otherwise.