Removing unused `@Nullable` on builder setters.
This change is removing `@Nullable` on arguments of builder setters for case where the getter was not annotated with `@Nullable`. This now results in an auto-value error, because in these case the setter was still performing a null check on the arguments, throwing an NPE if they were `null`. This lead to the erroneous assumption that `null` values could be passed in, when in fact they could not.