Remove permission check for Search Nav Item (#8917)
* Remove permission check for Search Nav Item

## Motivation
Prior to this change, the Search Menu Item was only displayd for admins
or users which have searches:absolute', 'searches:relative',
'searches:keyword' permissions.

The mentioned permission could only be add via API requests. Which left
the admin users as the only ones which could really see that Menu Item.

But a user could go to that page either via a link (on the streams page)
or by typing the url manually.

## Description
This change will now simply remove this check and always display the
menu item. The search still will only work if the user has read
permissions to a stream.

* Change default start page to search

the start page was set depending on the
search permissions a user has.

Now the start page will always default to Search Page