Show better error message when user has no stream permissions.
Before this change, the user was seeing a "Missing stream permissions"
page, that did not show a helpful error message if the user is not
missing the permission for a specific stream, but for any stream at all.

This change is now checking if the user is having any stream permissions
at all and shows a better message otherwise.

Fixes #8955.