Provide helpful error messages for ES misconfiguration/errors. (#8956)
Before this change, connecting against an unsupported ES node or failing
to probe the version of the running cluster resulted in a messy guice
exception. This change is throwing more specific exceptions in providers
and catching them in `CmdLineTool`, adding helpful error messages and
document links and bailing out in these case.