Fix path traversal in swagger resource.
Prior to this change, the `/api-browser/.*` resource allowed to retrieve
static assets used by swagger. Some of these were prefixed with a
`global/` prefix, that needs to be stripped away before locating the
file. The string replacement used for this was used unbounded (replacing
every occurence of the prefix, not just the first one), so it could be
used to create URLs containing `..`. The check if the route contains
`..` was performed _before_ the string replacement, so it could
effectively be used to evade it.

This change is:
  a) Replacing only the first occurrence of `global/`
  b) Performing the check if the route contains `..` _after_ the
replacement and _right before_ the file is tried to be opened.