Avoid URL synchronization if view is not currently dirty.
Before this change, the URL synchronization hook unconditionally added
query/timerange/streams information to the current URL if the current
view is a search (and not a dashboard). This lead to an unnecessary URL
change if a saved search is loaded (as the persisted
query/timerange/streams are identical to the current state).

This change is checking if the current view is dirty. If not, no change
is made.