Introduce EncryptedValue and related classes
This class contains an encrypted value and its salt. I can be used when
sensitive data must be stored in the database.

It comes with a custom JSON serializer and deserializer. They serialize
and deserialize the value in different ways depending on the context.
When storing/reading the value in/from MongoDB, the literal content
(encrypted value and salt) are read/written.
In all other contexts, the serialized JSON only contains an indicator if
the value is set or not. When parsing the JSON, only a "set_value" field
is allowed to set a new value.

The deserializer automatically creates an encrypted value and salt from
the given new plain text value.