Add authentication services infrastructure (#9007)
* Start identity provider authenticating realm

* Start a MongoDB identity provider

* Remove unused fields

* Add global identity provider config

* Start user profile and profile provisioner

* Better realm name

* Add comment

* Rename identity provider to authentication service

* Rename some fields

* Add AuthServiceBackend#backendType()

* Rename UserProfileProvisioner to ProvisionerService and add actions

* Add ProvisionException

* Add constant for default session timeout

* Add auth_service_id and auth_service_uid to User and UserService

* Provision a new user for external auth services

* Use a map binder and factories for provisioner actions

* Adjust ProvisionerService#newDetails()

* Set placeholder password for new users

* Introduce EncryptedValue and related classes

This class contains an encrypted value and its salt. I can be used when
sensitive data must be stored in the database.

It comes with a custom JSON serializer and deserializer. They serialize
and deserialize the value in different ways depending on the context.
When storing/reading the value in/from MongoDB, the literal content
(encrypted value and salt) are read/written.
In all other contexts, the serialized JSON only contains an indicator if
the value is set or not. When parsing the JSON, only a "set_value" field
is allowed to set a new value.

The deserializer automatically creates an encrypted value and salt from
the given new plain text value.

* Add UserService#loadAllForAuthServiceBackend

* Extend GlobalAuthServiceConfig for updates

* Add ValidationFailureException and corresponding mapper

* Implement auth backends and configurations

* Extract TrustManagerProvider into its own file so it can be reused

* Fix setting auth service attributes on User

* Add missing attribute to builder

* Fix user provisioning

We need to set an encrypted password value when creating a user.

* Include backend type in realm string in returned SimpleAccount

* Add missing user attributes to UserOverviewDTO

* Fall back to default backend if active backend doesn't work

* Remove unused variable

* Return null instead of a 404 if there is no active backend

* Add LdapEntry#getNonBlank and add some more tests

* Encrypt user password before passing it on to authentication services

This ensures that we don't expose it by accident (e.g. logging it)
All consumers need to decrypt it before they can use it.

* Reduce log level for some log statements

* Add auth service backend test code

* Add LDAP authentication service

* Don't run authentication services for the "root_username" user

* Use an explicit authenticating realm order instead of the dynamic one

Also disable binding of the legacy LDAP and MongoDB realm.s

* Fail login if provisioning fails

* Add TestPasswordSecretModule for guice injections in testing

* Add missing audit annotations to resources

* Add /system/authentication/services/active-backend/users endpoint