Submit widget query form before saving edit widget modal. (#8985)
* Submit widget query form before saving edit widget modal.

Before this change, a successful change to the timerange/stream
selection/query of a widget on a dashboard required submitting the
widget's search after the change and before clicking "Save". This was
unintuitive and unnecessarily complicated.

This change is now submitting the widget's query form when the user
clicks "Save" and the form is dirty, not requiring the user to have
submitted the form manually before.

Fixes #7922.

* Submit only if dirty.

* Removing unnecessary form element around events checkbox.

* Updating snapshots.

* Removing usages of `widget` prop for `WidgetQueryControls`.

* Reusing `DEFAULT_TIMERANGE`.

* Include `dirty` in `useCallback` dependencies.