Avoid unnecessary search execution when starting/loading search.
Before this change, starting a new or loading an existing search
triggered a search job execution at least twice. This leads to
unnecessary load and might even be the cause for bugs which are hard to
reproduce (#8820).

This change is removing two unneeded search executions during mount.
Both are factually unneeded, as the initial search execution after
loading the saved search is performed in `_refreshIfNotUndeclared`.